python_backend/requirements.txt
boto3>=1.34.0
botocore>=1.34.0
aws-lambda-powertools>=2.0.0
requests>=2.28.0
PyJWT>=2.8.0
bcrypt>=4.0.0
